refactor(Buy): move ExchangeService functionality into BuyDataManager
