+++ app/src/main/java/piuk/blockchain/android/data/datamanagers/BuyDataManager.java
@@ -1,34 +1,56 @@
 package piuk.blockchain.android.data.datamanagers;
 
+import android.util.Log;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import info.blockchain.wallet.metadata.Metadata;
+import info.blockchain.wallet.payload.PayloadManager;
+import info.blockchain.wallet.util.MetadataUtil;
+
+import org.bitcoinj.crypto.DeterministicKey;
+
+import java.util.ArrayList;
+import java.util.List;
+
 import io.reactivex.Observable;
-import piuk.blockchain.android.data.services.ExchangeService;
+import io.reactivex.subjects.ReplaySubject;
+import piuk.blockchain.android.data.exchange.ExchangeData;
+import piuk.blockchain.android.data.exchange.TradeData;
+import piuk.blockchain.android.data.rxjava.RxUtil;
 
 /**
  * Created by justin on 4/28/17.
  */
 
 public class BuyDataManager {
-    public static final String TAG = BuyDataManager.class.getSimpleName();
+    private static final String TAG = BuyDataManager.class.getSimpleName();
+    private static final int METADATA_TYPE_EXCHANGE = 3;
+    private static BuyDataManager instance;
 
     private OnboardingDataManager onboardingDataManager;
     private SettingsDataManager settingsDataManager;
     private PayloadDataManager payloadDataManager;
-    private ExchangeService exchangeService;
 
-    public BuyDataManager(OnboardingDataManager onboardingDataManager, SettingsDataManager settingsDataManager, PayloadDataManager payloadDataManager) {
+    private PayloadManager payloadManager;
+    private ReplaySubject<Metadata> metadataSubject;
+
+    private boolean didStartLoad = false;
+
+    private BuyDataManager(OnboardingDataManager onboardingDataManager, SettingsDataManager settingsDataManager, PayloadDataManager payloadDataManager) {
         this.onboardingDataManager = onboardingDataManager;
         this.settingsDataManager = settingsDataManager;
         this.payloadDataManager = payloadDataManager;
-        this.exchangeService = ExchangeService.getInstance();
+
+        this.payloadManager = PayloadManager.getInstance();
+        this.metadataSubject = ReplaySubject.create(1);
     }
 
-    public Observable<Metadata> getExchangeData() {
-        return this.exchangeService.getExchangeData();
-    }
-
-    public Observable<String> getPendingTradeAddresses() {
-        return exchangeService.getPendingTradeAddresses();
+    public static BuyDataManager getInstance(OnboardingDataManager onboardingDataManager, SettingsDataManager settingsDataManager, PayloadDataManager payloadDataManager) {
+        if (instance == null) {
+            instance = new BuyDataManager(onboardingDataManager, settingsDataManager, payloadDataManager);
+        }
+        return instance;
     }
 
     public Observable<Boolean> getCanBuy() {
@@ -49,7 +71,62 @@ public class BuyDataManager {
         });
     }
 
+    public Observable<String> getPendingTradeAddresses() {
+        Log.d(TAG, "getPendingTradeAddresses: called");
+        return this.getExchangeData()
+                .flatMap(metadata ->
+                        Observable.fromCallable(metadata::getMetadata)
+                )
+                .flatMapIterable(exchangeData -> {
+                    ObjectMapper mapper = new ObjectMapper();
+                    ExchangeData data = mapper.readValue(exchangeData, ExchangeData.class);
+
+                    List<TradeData> trades = new ArrayList<TradeData>();
+                    if (data.getCoinify() != null) {
+                        trades.addAll(data.getCoinify().getTrades());
+                    } else if (data.getSfox() != null) {
+                        trades.addAll(data.getSfox().getTrades());
+                    }
+
+                    return trades;
+                })
+                .filter(tradeData ->
+                        !tradeData.isConfirmed()
+                )
+                .map(tradeData ->
+                        payloadDataManager.getReceiveAddressAtPosition(
+                                payloadDataManager.getAccount(tradeData.getAccountIndex()),
+                                tradeData.getReceiveIndex()
+                        )
+                )
+                .map(address -> {
+                    Log.d(TAG, "getPendingTradeAddresses: found address: " + address);
+                    return address;
+                })
+                .distinct();
+    }
+
+    public Observable<Metadata> getExchangeData() {
+        if (!didStartLoad) {
+            reloadExchangeData();
+            didStartLoad = true;
+        }
+        return metadataSubject;
+    }
+
     public void reloadExchangeData() {
-        this.exchangeService.reloadExchangeData();
+        Observable<Metadata> exchangeDataStream = getMetadata();
+        exchangeDataStream.subscribeWith(metadataSubject);
+    }
+
+    private Observable<Metadata> getMetadata() {
+        return Observable.fromCallable(() -> {
+            DeterministicKey masterKey = payloadManager
+                    .getPayload()
+                    .getHdWallets().get(0)
+                    .getMasterKey();
+            DeterministicKey metadataHDNode = MetadataUtil.deriveMetadataNode(masterKey);
+            return new Metadata.Builder(metadataHDNode, METADATA_TYPE_EXCHANGE).build();
+        }).compose(RxUtil.applySchedulersToObservable());
     }
 }
